Move MockEvaluator from early stopping to mocks module
