+++ tests/mocks.py
@@ -2,14 +2,17 @@
 
 """Mocks for tests."""
 
-from typing import Optional, Tuple
+from typing import Iterable, Optional, Tuple
 
 import torch
 from torch import nn
 
+from pykeen.evaluation import Evaluator, MetricResults, RankBasedMetricResults
+from pykeen.evaluation.rank_based_evaluator import RANK_REALISTIC, RANK_TYPES, SIDES
 from pykeen.models import EntityRelationEmbeddingModel, Model
 from pykeen.nn.emb import EmbeddingSpecification, RepresentationModule
 from pykeen.triples import CoreTriplesFactory
+from pykeen.typing import MappedTriples
 
 __all__ = [
     'CustomRepresentations',
@@ -59,3 +62,78 @@ class MockModel(EntityRelationEmbeddingModel):
 
     def reset_parameters_(self) -> Model:  # noqa: D102
         pass  # Not needed for unittest
+
+
+class MockEvaluator(Evaluator):
+    """A mock evaluator for testing early stopping."""
+
+    def __init__(self, losses: Iterable[float], automatic_memory_optimization: bool = True) -> None:
+        super().__init__(automatic_memory_optimization=automatic_memory_optimization)
+        self.losses = tuple(losses)
+        self.losses_iter = iter(self.losses)
+
+    def process_tail_scores_(
+        self,
+        hrt_batch: MappedTriples,
+        true_scores: torch.FloatTensor,
+        scores: torch.FloatTensor,
+        dense_positive_mask: Optional[torch.FloatTensor] = None,
+    ) -> None:  # noqa: D102
+        pass
+
+    def process_head_scores_(
+        self,
+        hrt_batch: MappedTriples,
+        true_scores: torch.FloatTensor,
+        scores: torch.FloatTensor,
+        dense_positive_mask: Optional[torch.FloatTensor] = None,
+    ) -> None:  # noqa: D102
+        pass
+
+    def finalize(self) -> MetricResults:  # noqa: D102
+        hits = next(self.losses_iter)
+        dummy_1 = {
+            side: {
+                rank_type: 10
+                for rank_type in RANK_TYPES
+            }
+            for side in SIDES
+        }
+        dummy_2 = {
+            side: {
+                rank_type: 1.0
+                for rank_type in RANK_TYPES
+            }
+            for side in SIDES
+        }
+        return RankBasedMetricResults(
+            arithmetic_mean_rank=dummy_1,
+            geometric_mean_rank=dummy_1,
+            harmonic_mean_rank=dummy_1,
+            median_rank=dummy_1,
+            inverse_arithmetic_mean_rank=dummy_2,
+            inverse_harmonic_mean_rank=dummy_2,
+            inverse_geometric_mean_rank=dummy_2,
+            inverse_median_rank=dummy_2,
+            adjusted_arithmetic_mean_rank=dummy_2,
+            adjusted_arithmetic_mean_rank_index={
+                side: {
+                    RANK_REALISTIC: 0.0,
+                }
+                for side in SIDES
+            },
+            rank_std=dummy_1,
+            rank_var=dummy_1,
+            rank_mad=dummy_1,
+            hits_at_k={
+                side: {
+                    rank_type: {
+                        10: hits,
+                    } for rank_type in RANK_TYPES
+                }
+                for side in SIDES
+            },
+        )
+
+    def __repr__(self):  # noqa: D105
+        return f'{self.__class__.__name__}(losses={self.losses})'
